+++ components/defineComponents.ts
@@ -1,19 +1,3 @@
-/* Core CSS required for Ionic components to work properly */
-import "@ionic/core/css/core.css";
-
-/* Basic CSS for apps built with Ionic */
-import "@ionic/core/css/normalize.css";
-import "@ionic/core/css/structure.css";
-import "@ionic/core/css/typography.css";
-
-/* Optional CSS utils that can be commented out */
-import "@ionic/core/css/padding.css";
-import "@ionic/core/css/float-elements.css";
-import "@ionic/core/css/text-alignment.css";
-import "@ionic/core/css/text-transformation.css";
-import "@ionic/core/css/flex-utils.css";
-import "@ionic/core/css/display.css";
-
 /* All Ionic components */
 import { IonAccordion } from "@ionic/core/components/ion-accordion";
 import { IonAccordionGroup } from "@ionic/core/components/ion-accordion-group";
@@ -115,7 +99,10 @@ export const defineComponent = (tagName: string, customElement: any) => {
   }
 };
 
+
+
 export const defineIonComponents = () => {
+
   defineComponent("ion-accordion", IonAccordion);
   defineComponent("ion-accordion-group", IonAccordionGroup);
   defineComponent("ion-app", IonApp);