+++ CHANGELOG.md
@@ -4,6 +4,62 @@ All notable changes to this project will be documented in this file.
 
 ## 0.5.30
 - Changed the default starter to use the cjs import, also to enable codesplitting - creator 0.0.8
+- Config routine changed for base `+layout.svelte`
+
+```
+<script lang="ts">
+	import { setupIonicSvelte } from 'ionic-svelte';
+
+	/* Theme variables */
+	import '../theme/variables.css';
+
+	/* Call Ionic's setup routine */
+	setupIonicSvelte();
+
+</script>
+
+<ion-app>
+	<slot />
+</ion-app>
+
+```
+Should become:
+```
+<script lang="ts">
+	import { setupIonicBase } from 'ionic-svelte';
+
+	/* Call Ionic's setup routine */
+	setupIonicBase();
+
+	/* Theme variables */
+	import '../theme/variables.css';
+
+	/*
+		The next command loads and registers all Ionic Webcomponents for you to use.
+
+		This adds at least >800kb (uncompressed) to your bundle.
+
+		You can also choose to import each component you want to use separately.
+
+		It is recommended to do this in this file, as you only need to do such once. But you are free
+		to do this elsewhere if you like to code-split differently.
+
+		You can find these import lines to copy here - https://github.com/Tommertom/svelte-ionic-npm/blob/main/components/all.js
+	
+		Please don't forget to import ion-app in this file when you decide to code-split:
+		
+	    import 'ionic-svelte/components/ion-app';
+	*/
+	import 'ionic-svelte/components/all';
+</script>
+
+<ion-app>
+	<slot />
+</ion-app>
+
+
+```
+
 
 ## 0.5.29
 - Updated typings